Add files via upload
This project is focused on performing Exploratory Data Analysis (EDA) and Data Preprocessing on a loan dataset. The goal is to clean and preprocess the data, and provide insights into trends and relationships between different features of the loan data. The project follows a step-by-step approach to data exploration, cleaning, and transformation.